updating run script to be modular
