+++ cli/lib/write-seed-env-file.ts
@@ -4,10 +4,8 @@ import { fileURLToPath } from "url";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
-const configFilePath = path.resolve(
-  path.join(__dirname, "../../../"),
-  ".env.seed"
-);
+const outputPath = path.join(__dirname, "../../");
+const configFilePath = path.resolve(outputPath, ".env.seed");
 
 export const writeSeedEnvFile = async (
   envVariables: Record<string, string>
