+++ deployment/stacks/env-config.template.js
@@ -1,7 +1,6 @@
 window._env_ = {
   API_REGION: "us-east-1",
   API_URL: "{{apiGatewayRestApiUrl}}",
-  APPLICATION_ENDPOINT: "{{applicationEndpointUrl}}",
   COGNITO_IDENTITY_POOL_ID: "{{identityPoolId}}",
   COGNITO_IDP_NAME: "Okta",
   COGNITO_REDIRECT_SIGNIN: "{{applicationEndpointUrl}}",
@@ -10,8 +9,6 @@ window._env_ = {
   COGNITO_USER_POOL_CLIENT_DOMAIN: "{{userPoolClientDomain}}",
   COGNITO_USER_POOL_CLIENT_ID: "{{userPoolClientId}}",
   COGNITO_USER_POOL_ID: "{{userPoolId}}",
-  LD_PROJECT_KEY: "mdct-mcr",
-  LOCAL_LOGIN: "false",
   POST_SIGNOUT_REDIRECT: "{{redirectSignout}}",
   REACT_APP_LD_SDK_CLIENT: "{{launchDarklyClient}}",
   SKIP_PREFLIGHT_CHECK: "true",
