--- src/write-ui-env-file.ts
@@ -1,45 +1,0 @@
-import path, { dirname } from "path";
-import { promises as fs } from "fs";
-import { fileURLToPath } from "url";
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = dirname(__filename);
-const outputPath = path.join(__dirname, "../../services/ui-src", "public");
-const configFilePath = path.resolve(outputPath, "env-config.js");
-
-const region = "us-east-1";
-
-export async function writeLocalUiEnvFile(apiUrl: string) {
-  const envVariables = {
-    API_REGION: region,
-    API_URL: apiUrl.replace("https", "http"),
-    APPLICATION_ENDPOINT: "http://localhost:3000/",
-    COGNITO_IDENTITY_POOL_ID: process.env.COGNITO_IDENTITY_POOL_ID,
-    COGNITO_IDP_NAME: "Okta",
-    COGNITO_REDIRECT_SIGNIN: "http://localhost:3000/",
-    COGNITO_REDIRECT_SIGNOUT: "http://localhost:3000/postLogout",
-    COGNITO_REGION: region,
-    COGNITO_USER_POOL_CLIENT_DOMAIN:
-      process.env.COGNITO_USER_POOL_CLIENT_DOMAIN,
-    COGNITO_USER_POOL_CLIENT_ID: process.env.COGNITO_USER_POOL_CLIENT_ID,
-    COGNITO_USER_POOL_ID: process.env.COGNITO_USER_POOL_ID,
-    LD_PROJECT_KEY: "mdct-mcr",
-    LOCAL_LOGIN: "false",
-    POST_SIGNOUT_REDIRECT: "http://localhost:3000/",
-    REACT_APP_LD_SDK_CLIENT: process.env.REACT_APP_LD_SDK_CLIENT,
-    SKIP_PREFLIGHT_CHECK: "true",
-    STAGE: "local",
-  };
-
-  await fs.rm(configFilePath, { force: true });
-
-  const envConfigContent = [
-    "window._env_ = {",
-    ...Object.entries(envVariables).map(
-      ([key, value]) => `  ${key}: "${value}",`
-    ),
-    "};",
-  ].join("\n");
-
-  await fs.writeFile(configFilePath, envConfigContent);
-}
